+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.4.0 (unreleased)
 ------------------
 
+- #2239 Allow to create multiple samples for each sample record in add form
 - #2241 Little improvement of getRaw function from legacy uidreference field
 - #2238 Split the add sample's `ajax_form` function to make patching easier
 - #2240 Explicitly set client on sample creation